refactor: actions, selectors, reducers
