+++ src/store/actions/index.ts
@@ -1,0 +1,3 @@
+export { userActions } from './userActions';
+export { addGameSettings } from './addGameSettings';
+export { messageActions } from './messageActions';
