+++ src/store/actions/messageActions.ts
@@ -1,5 +1,7 @@
 import { createAction } from '@reduxjs/toolkit';
 import { IMessage } from '@/utils/interfaces';
 
-export const addMessage = createAction<IMessage>('addMessage');
-export const getMessages = createAction<IMessage[]>('getMessages');
+export const messageActions = {
+  addMessage: createAction<IMessage>('addMessage'),
+  getMessages: createAction<IMessage[]>('getMessages'),
+};
