+++ src/store/reducers/index.ts
@@ -1,0 +1,3 @@
+export { gameSettings } from './gameSettingsReducer';
+export { messages } from './messagesReducer';
+export { user } from './userReducer';
