+++ src/store/selectors/index.ts
@@ -1,3 +1,3 @@
 export { gameSettingsSelector } from './gameSettingsSelector';
-export { messageSelector } from './messageSelector';
+export { messagesSelectors } from './messageSelector';
 export { userSelectors } from './userSelectors';
