+++ src/store/selectors/messageSelector.ts
@@ -1,4 +1,6 @@
 import { createSelector } from '@reduxjs/toolkit';
 import { RootState } from '@/store';
 
-export const messageSelector = createSelector((state: RootState) => state.messages, (messages) => messages);
+export const messagesSelectors = {
+  messageSelector: createSelector((state: RootState) => state.messages, (messages) => messages),
+};
